refactor(status): Create a `status` module
The corresponding code for `bt status` is moved under its own module to
keep lib root lean and achieve modularity within `bt`.
