+++ src/lib.rs
@@ -1,35 +1,8 @@
 pub mod api;
-
 use std::fmt::Debug;
 use std::{error, fmt, io};
 use tabled::builder::Builder;
 use tabled::settings::Style;
-pub fn status(f: &mut impl io::Write) -> Result<(), Box<dyn error::Error>> {
-    let bluez = Bluez::new()?;
-
-    let power_state = bluez.power_state()?;
-    let connected_devs = bluez.connected_devs()?;
-
-    let mut buf = [
-        "bluetooth: ",
-        &power_state.to_string(),
-        "\nconnected devices: ",
-    ]
-    .join("");
-    for dev in connected_devs {
-        let format = format!(
-            "\n{}/{} (batt: %{})",
-            dev.alias,
-            dev.address,
-            dev.battery.unwrap()
-        );
-        buf.push_str(&format)
-    }
-
-    f.write_all(buf.as_bytes())?;
-
-    Ok(())
-}
 
 pub fn toggle(f: &mut impl io::Write) -> Result<(), Box<dyn error::Error>> {
     let bluez = Bluez::new()?;
@@ -171,3 +144,6 @@ pub fn create_terse_out(listing: impl Iterator<Item = Vec<String>>) -> String {
         .collect()
 }
 mod bluez;
+mod status;
+
+pub use status::status;
